requirements.txt
ultralytics>=8.3.0
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
pytest>=7.0.0 